Update managed disk docs with correct size types, remove import tests & tidy tests
