+++ website/docs/r/managed_disk.html.markdown
@@ -82,7 +82,7 @@ The following arguments are supported:
     Changing this forces a new resource to be created.
 
 * `storage_account_type` - (Required) The type of storage to use for the managed disk.
-    Allowable values are `Standard_LRS` or `Premium_LRS`.
+    Allowable values are `Standard_LRS`, `Premium_LRS`, `StandardSSD_LRS` or `UltraSSD_LRS`.
 
 * `create_option` - (Required) The method to use when creating the managed disk. Possible values include:
  * `Import` - Import a VHD file in to the managed disk (VHD specified with `source_uri`).
